Fixed recursive Objects
If an object A has a reference back to itself (ie A.a = this) then
previously this would give stack overflows or hang. This new version
works correctly, but only if batchLoad = true (default). If BatchLoad =
false, then the resolution of the child objects occurs before the
creation of this object, not allowing for a proper replacement strategy.
This has to be the case for constructors to work. Honestly, there is
very look good reason not to use BatchLoad so maybe this should be
deprecated.