Implemented factory constructors
Added the ability to construct an object from a factory method. This is
useful in situations where the underlying objects are created from
external sources, for example when mapped to ProtoBuf objects via
builder methods.

As part of this change, relaxed some of the validation on setters, such
as the setter cannot return anything other than void. This allows for
setters which implement method chaining to be used.

Also allowed for null AerospikeClients to be passed in. This is useful
in situations where object mapping is desired via the mapping converter
but does not need the data to be saved into Aerospike.