Added the ability to use getters / setters
Added a parameter to AerospikeBin which forces the mapper to use the
getters/setters for a field instead of directly setting the field. This
is useful when the getters/setters have logic in them beyond simply
setting the attribute. This could have been done previously by excluding
the field and marking the getters and setters explicitly, but this was
clunky.