Allowed superclasses to have no attributes
An error message was thrown if a class had no immediate or superclass
attributes to be saved to the database. This is a valid check on a low
level class, but it's entirely fair to have a superclass (typically
abstract) which has no attributes, and this error was being thrown here
too.

The check is safe to remove as it's illegal to save a class with no
attributes to the database because there will be no @Id field so will
throw an exception anyway.