Allowed the top of the hierarchy to be an interface.
If classes implement an interface instead of extending an abstract
class, the Java Object Mapper should respect any AerospikeRecords
annotations on the interface.

The Object Mapper will look for definitions in the config file, then any
AerospikeRecords on the specific class, then the superclasses and
finally anything in the interface hierarchy. The first one which defines
a namespace will be the definition which is used.